fixing samples
