+++ Samples/CMakeLists.txt
@@ -1,5 +1,3 @@
-#add_subdirectory(Utils)
-
 add_subdirectory(Tcp)
 
-add_subdirectory(Udp)
+# add_subdirectory(Udp)
