+++ Samples/ProcessLauncher.h
@@ -5,24 +5,23 @@
 * report any bug to andrecasa91@gmail.com.
 **/
 
-#ifndef _PROCESS_LAUNCHER_H_
-#define _PROCESS_LAUNCHER_H_
+#ifndef SAMPLE_LAUNCHER_H
+#define SAMPLE_LAUNCHER_H
 
-#include <fstream>
 #include <string>
 #include <list>
 #include <utility>
 
 /**    
  * @brief The object is used to print a script, that is able
- * to launch in a precise sequence certain processes, opening different prompts
+ * to launch in a precise sequence certain processes
  */
 class Launcher {
 public:
     /**    
      * @param[in] the name of script to print
      */
-   Launcher(const std::string& nameFile);
+   Launcher(const std::string& nameFile) : nameFile(nameFile) {};
 
     /** 
      * @brief Add 1 process to launch, with possibly some additional command arguments.
@@ -43,6 +42,7 @@ public:
    void  operator()() const;
 
 private:
+
    template <typename ... Args>
    void parseArgs(std::list<std::string>& parsed , const std::string& arg, Args ... args) {
       parsed.push_back(arg);
@@ -60,6 +60,47 @@ private:
 
    std::string nameFile;
    std::list<std::pair<std::string, std::list<std::string>>> commands;
+};
+
+#include <fstream>
+
+void  Launcher::operator()() const {
+    std::string name = this->nameFile;
+#ifdef _WIN32
+    name += ".bat";
+#elif  __linux__
+    name += ".sh";
+#endif
+
+    std::ofstream f(name);
+    auto it = this->commands.begin();
+    auto itEnd = this->commands.end();
+    --itEnd;
+    for (it; it != itEnd; ++it) {
+#ifdef _WIN32
+        f << std::endl << "start \"\" \"" << it->first << "\"";
+        for(auto a : it->second) f << " \"" << a << "\"";
+#elif  __linux__
+        f << std::endl << "gnome-terminal -x sh -c \"./" << it->first;
+        for(auto a : it->second) f << " " << a;
+        f << "; bash\"";
+#endif
+    }
+#ifdef _WIN32
+    f << std::endl << "\"" << this->commands.back().first << "\""; 
+    for(auto a : this->commands.back().second) f << " \"" << a << "\"";
+#elif  __linux__
+    f << std::endl << "./" << this->commands.back().first;
+    for(auto a : this->commands.back().second) f << " " << a;
+#endif
+
+    f.close();
+
+#ifdef _WIN32
+    system(name.c_str());
+#elif  __linux__
+    system(std::string("sh ./" + name).c_str());
+#endif
 };
 
 #endif
