+++ Samples/Tcp/01-server-client/Launcher.cpp
@@ -1,0 +1,11 @@
+#include <ProcessLauncher.h>
+
+int main() {
+
+   Launcher lnc("launcher");
+   lnc.addProcess("Sample01_Server");
+   lnc.addProcess("Sample01_Client");
+   lnc();
+
+   return 0;
+}
