+++ Samples/Tcp/CMakeLists.txt
@@ -1,7 +1,7 @@
-add_subdirectory(Sample_01_server_client)
+add_subdirectory(01-server-client)
 
-add_subdirectory(Sample_02_singlethread_services)
+# add_subdirectory(Sample_02_singlethread_services)
 
-add_subdirectory(Sample_03_multithread_services)
+# add_subdirectory(Sample_03_multithread_services)
 
-add_subdirectory(Sample_04_repeater)
+# add_subdirectory(Sample_04_repeater)
