--- Samples/Tcp/Sample_01_server_client/Launcher.cpp
@@ -1,11 +1,0 @@
-#include <ProcessLauncher.h>
-
-int main() {
-
-   Launcher lnc("launcher");
-   lnc.addProcess("Sample01_Server");
-   lnc.addProcess("Sample01_Client");
-   lnc();
-
-   return 0;
-}
