--- Samples/Tcp/Sample_04_repeater/Launcher.cpp
@@ -1,12 +1,0 @@
-#include <ProcessLauncher.h>
-
-int main() {
-
-   Launcher lnc("launcher");
-   lnc.addProcess("Sample04_Server");
-   lnc.addProcess("Sample04_Repeater");
-   lnc.addProcess("Sample04_Client");
-   lnc();
-
-   return 0;
-}
