--- Samples/Utils/CMakeLists.txt
@@ -1,22 +1,0 @@
-set(PROJECT_NAME "Utils")
-
-CollectSourceFiles(${CMAKE_CURRENT_SOURCE_DIR} SOURCES)
-
-GroupSources(${CMAKE_CURRENT_SOURCE_DIR})
-
-if(LIB_OPT)
-	add_library(${PROJECT_NAME} SHARED ${SOURCES})
-else()
-	add_library(${PROJECT_NAME} STATIC ${SOURCES})
-endif()
-
-target_include_directories(${PROJECT_NAME}
-    PUBLIC
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-        $<INSTALL_INTERFACE:include>
-)
-
-target_link_libraries(${PROJECT_NAME}
-    PUBLIC
-	SCK::Cross-Socket
-)
