--- cmake/CheckPlatform.cmake
@@ -1,10 +1,0 @@
-macro(check_platform platform)
-# check what platform we're on (64-bit or 32-bit), and create a simpler test than CMAKE_SIZEOF_VOID_P
-if(CMAKE_SIZEOF_VOID_P MATCHES 8)
-    set(${platform} 64)
-    #MESSAGE(STATUS "Detected 64-bit platform")
-else()
-    set(${platform} 32)
-    #MESSAGE(STATUS "Detected 32-bit platform")
-endif()
-endmacro(check_platform)
