--- cmake/GetCompiler.cmake
@@ -1,13 +1,0 @@
-macro(get_compiler compiler)
-  if ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "Clang")
-    set(${compiler} "clang")
-  elseif ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "GNU")
-    set(${compiler} "gcc")
-  elseif ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "Intel")
-    set(${compiler} "intel")
-  elseif ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "MSVC")
-    set(${compiler} "msvc")
-  else()
-    set(${compiler} "unknown")
-  endif()
-endmacro(get_compiler)
