--- cmake/Log.cmake
@@ -1,8 +1,0 @@
-function(debug msg line)
-    if(DEBUG)
-      message(STATUS "[ ${CMAKE_CURRENT_LIST_FILE}:${line} ] " ${msg})
-    endif()
-endfunction(debug)
-function(log msg line)
-    message(STATUS "[ ${CMAKE_CURRENT_LIST_FILE}:${line} ] " ${msg})
-endfunction(log)
